+++ source/SIMULATION/PTMSimulation.C
@@ -120,7 +120,7 @@ namespace OpenMS {
   
 		if (ptms_.size()==0) return;
 
-		Size max_mod_count = (UInt) param_.getValue("modification_bound");
+		Size max_mod_count = param_.getValue("modification_bound");
 		if (max_mod_count == 0) return;
 
 		FeatureMapSim map_ptm;
